Fixed some CSS -- Added Wrapper -- Clean up
Check out the revision notes in GitHub for Mac for this checkin. I
added a wrapper in to contain the whole site. Also be sure to pay
attention to what styles go where:

Anything for TABLET or MOBILE, widths should be in percentages so the
divs remain fluid. Desktop is ok to get pixel values.

Also remember anything at the very top of the stylesheet is for Mobile
and global css elements. Use the MediaQueries sections in the CSS to
perform custom css enhancements for larger and larger screens. Mobile
is first.

You’ll notice it is now breaking down responsively, and you’ll see
percentages in the top sections and when you get to the desktop (1024)
media query values move to PX values.

Take another stab from here and see what you can come up with.

Best - MW